Modify HTML, CSS, and JS
Modified index.html to add the UI for the calculator as well as header and footers.
Modified style.css to add styling for the entire page. Made the calculator look nice.
Modified script.js but the function to get the first number, second number, and the operator is not finished and is going to change.