Modify HTML and JS files
Modified index.html to add a backspace button and an onclick function
Modified script.js to add functionality to slice the last character of the current input string so that the backspace button functions

This is my last commit for the forseeable future. Everything pretty much works as intended and all of the extra credit section is done besides keyboard support (I may add this later).
For now, I am done with this project.